--- OZprivate/rawJS/OZTreeModule/src/render/line_to_renderer.js
@@ -1,5 +1,0 @@
-function follow_path(context, shape) {
-  context.lineTo(shape.x, shape.y);
-}
-
-export {follow_path};
