--- OZprivate/rawJS/OZTreeModule/src/render/move_to_renderer.js
@@ -1,5 +1,0 @@
-function follow_path(context, shape) {
-  context.moveTo(shape.x, shape.y);
-}
-
-export {follow_path};
